Add awareness of Python specifiers in project files for autodetection
Certain files, namely `pyproject.toml` and `runtime.txt`, can be used to
specify which python versions your project is compatible with.

Since the parsing logic for python specifiers and detecting these files
can be a bit hairy, we implement this in Python and trust the user to
have some sort of basic Python 3 environment available before installing
this pyenv plugin.

We also skip full compliance with the entire [python version-specifier spec],
since it's too long and we can probably make do with a basic subset for
now.

Fixes: https://github.com/strokirk/pyenv-autoenv/issues/1
Fixes: https://github.com/strokirk/pyenv-autoenv/issues/2

[python version-specifier spec]: https://www.python.org/dev/peps/pep-0440/#version-specifiers